// max-most-frontend/src/Interfaces/Company.ts
export interface Website {
  authorization_key?: string;
  consumer_key?: string | null;
  consumer_secret?: string | null;
  id: string;
  site_url: string;
  title: string;
  value?: string;
  label?: string;
  is_trash?: boolean | null;
}